Appears to match functionality of polyskel (#1)
Not all tests pass, but that's because the subtrees that the two
versions (cpp and python original) are slightly different in a few
examples.

This is because the python prioqueue does not keep insertion order when
priorities are the same. So it's kind of impossible to replicate its
behaviour in cpp.

Visually, the examples seem to match.