Requirements.txt (#2)
* Appears to match functionality of polyskel

Not all tests pass, but that's because the subtrees that the two
versions (cpp and python original) are slightly different in a few
examples.

This is because the python prioqueue does not keep insertion order when
priorities are the same. So it's kind of impossible to replicate its
behaviour in cpp.

Visually, the examples seem to match.

* Preview text, requirements.txt for tests

Preview images now show text (cpp or python) so that you can tell which
one came from cpp and which from python.

Also added requirements.txt for pillow in tests.